feat: Adicionada verificação se arquivo já está presente no destino final. Ignora caso presente

+++ organizer.py
@@ -2,7 +2,7 @@ import os
 import shutil
 
 path = input('Insert path to organize: ')
-
+existing_files = 0
 files = os.listdir(path)
 try:
     for file in files:
@@ -10,10 +10,18 @@ try:
         extension = extension[1:]
 
         if os.path.exists(path+'/'+extension):
-            shutil.move(path+'/'+file, path+'/'+extension+'/'+file)
+            if not os.path.exists(path+'/'+extension+'/'+file):
+                shutil.move(path+'/'+file, path+'/'+extension+'/'+file)
+            else:
+                existing_files += 1
         else:
             os.makedirs(path+'/'+extension)
-            shutil.move(path+'/'+file, path+'/'+extension+'/'+file)
+            if not os.path.exists(path+'/'+extension+'/'+file):
+                shutil.move(path+'/'+file, path+'/'+extension+'/'+file)
+            else:
+                existing_files += 1
+    if existing_files != 0:
+        print(f'Arquivos organizados com sucesso! {existing_files} arquivos já existentes não foram movidos.')
     print('Arquivos organizados com sucesso!')
 except Exception as e:
     print('Erro ao organizar arquivos: ', e)
